add devices endpoint

+++ src/main/java/com/mbb/api/sales_tracker/repository/BrandRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mbb.api.sales_tracker.model.Brand;
 
-public interface BrandRepository extends JpaRepository<Brand, Long> {
+public interface BrandRepository extends JpaRepository<Brand, Integer> {
 
 }
